// apps/web/src/types/codegen/api.d.ts
export interface components {
  schemas: {
// ...
    MediaType: {
      parameters: {
        [key: string]: string
      }
      subtype: string
      type: string
      wildcardSubtype: boolean
      wildcardType: boolean
    }
// ...
    NewCookie: {
      comment: string
      domain: string
      /** Format: date-time */
      expiry: string
      httpOnly: boolean
      /** Format: int32 */
      maxAge: number
      name: string
      path: string
      /** @enum {string} */
      sameSite: 'NONE' | 'LAX' | 'STRICT'
      secure: boolean
      value: string
      /** Format: int32 */
      version: number
    }
// ...
  }
  responses: never
  parameters: never
  requestBodies: never
  headers: never
  pathItems: never
}
export type $defs = Record<string, never>
export interface operations {
  getEnv: {
    parameters: {
      query?: never
      header?: never
      path?: never
      cookie?: never
    }
    requestBody?: never
    responses: {
      /** @description OK */
      200: {
        headers: {
          [name: string]: unknown
        }
        content: {
          '*/*': components['schemas']['Configuration']
        }
      }
    }
  }
  getAllUserConversations: {
    parameters: {
      query?: never
      header?: never
      path?: never
      cookie?: never
    }
    requestBody?: never
    responses: {
      /** @description OK */
      200: {
        headers: {
          [name: string]: unknown
        }
        content: {
          '*/*': components['schemas']['ConversationDTO'][]
        }
      }
    }
  }
  getSingleConversation: {
    parameters: {
      query?: never
      header?: never
      path: {
        conversationId: string
      }
      cookie?: never
    }
    requestBody?: never
    responses: {
      /** @description OK */
      200: {
        headers: {
          [name: string]: unknown
        }
        content: {
          '*/*': components['schemas']['ConversationDTO']
        }
      }
    }
  }
  deleteConversation: {
    parameters: {
      query?: never
      header?: never
      path: {
        conversationId: string
      }
      cookie?: never
    }
    requestBody?: never
    responses: {
      /** @description OK */
      200: {
        headers: {
          [name: string]: unknown
        }
        content?: never
      }
    }
  }
  getConversationHistory: {
    parameters: {
      query?: never
      header?: never
      path: {
        conversationId: string
      }
      cookie?: never
    }
    requestBody?: never
    responses: {
      /** @description OK */
      200: {
        headers: {
          [name: string]: unknown
        }
        content: {
          '*/*': components['schemas']['MessageDTO'][]
        }
      }
    }
  }
  streamGeneration: {
    parameters: {
      query?: never
      header?: never
      path: {
        conversationId: string
      }
      cookie?: never
    }
    requestBody: {
      content: {
        'application/json': components['schemas']['LLMRequest']
      }
    }
    responses: {
      /** @description OK */
      200: {
        headers: {
          [name: string]: unknown
        }
        content: {
          'text/event-stream': components['schemas']['ChatStreamResponseDTO'][]
        }
      }
    }
  }
  newSubscriptionWithTrial: {
    parameters: {
      query?: never
      header: {
        'X-priceId': string
      }
      path?: never
      cookie?: never
    }
    requestBody?: never
    responses: {
      /** @description OK */
      200: {
        headers: {
          [name: string]: unknown
        }
        content: {
          '*/*': string
        }
      }
    }
  }
  completeOnboarding: {
    parameters: {
      query?: never
      header?: never
      path?: never
      cookie?: never
    }
    requestBody?: never
    responses: {
      /** @description OK */
      200: {
        headers: {
          [name: string]: unknown
        }
        content: {
          '*/*': components['schemas']['Response']
        }
      }
    }
  }
  getUser: {
    parameters: {
      query?: never
      header?: never
      path?: never
      cookie?: never
    }
    requestBody?: never
    responses: {
      /** @description OK */
      200: {
        headers: {
          [name: string]: unknown
        }
        content: {
          '*/*': components['schemas']['UserDTO']
        }
      }
    }
  }
  updateUser: {
    parameters: {
      query: {
        user: components['schemas']['UserDTO']
      }
      header?: never
      path?: never
      cookie?: never
    }
    requestBody?: never
    responses: {
      /** @description OK */
      200: {
        headers: {
          [name: string]: unknown
        }
        content: {
          '*/*': components['schemas']['Response']
        }
      }
    }
  }
  handleStripeWebhook: {
    parameters: {
      query?: never
      header: {
        'Stripe-Signature': string
      }
      path?: never
      cookie?: never
    }
    requestBody: {
      content: {
        'application/json': string
      }
    }
    responses: {
      /** @description OK */
      200: {
        headers: {
          [name: string]: unknown
        }
        content: {
          '*/*': string
        }
      }
    }
  }
}